modernize more for conan v2
