+++ recipes/libsvm/all/test_package/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.1)
-project(PackageTest CXX)
+project(test_package LANGUAGES C)
 
 find_package(libsvm CONFIG REQUIRED)
 
-add_executable(test_package test_package.cpp)
+add_executable(test_package test_package.c)
 target_link_libraries(test_package PRIVATE libsvm::libsvm)
