--- recipes/libsvm/all/test_package/test_package.cpp
@@ -1,17 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include "svm/svm.h"
-
-struct svm_parameter param;
-
-int main(int argc, char **argv)
-{
-	param.svm_type = C_SVC;
-	param.kernel_type = PRECOMPUTED;
-
-	//Allocate some dummy data
-	param.weight = (double*)malloc(10 * sizeof(double));
-	svm_destroy_param(&param);
-
-	printf("libsvm version %d test_package OK \n", LIBSVM_VERSION);
-}
